rock-inspect: big performance improvement
rock-inspect was calling Inspect.find with an array while the call
expected a string. The call was then doing /#{pattern}/, which in
the case of an array is

  /["array1","array2"]/

in which the content of [] is interpreted as a character set and,
in practice, was matching everything as soon as one added ':' to
the search. Fix the call to have a proper composite pattern, which
reduces the needed time to search tremendously